Refactors to put service control classes in their own common module
